scraping/pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "openai>=1.52.0",
    "python-dotenv>=1.0.1",
]
